format filepath for WSL
